update NET_FLOW_STATISTICS function and fix bug (no grouping), Y190725

+++ CMAP/src/com/cmap/plugin/module/netflow/statistics/NetFlowStatisticsService.java
@@ -18,7 +18,7 @@ public interface NetFlowStatisticsService {
      * @throws ServiceLayerException
      */
     public void calculateIpTrafficStatistics(
-            String groupId, Date statDate, Map<String, Map<String, Integer>> ipTrafficMap) throws ServiceLayerException;
+            String groupId, Date statDate, Map<String, Map<String, Long>> ipTrafficMap) throws ServiceLayerException;
 
     /**
      * 取得 NET_FLOW 流量統計資料筆數
